alterando variavel de ambiente para 127.0.0.1

+++ src/controllers/blogPostController.js
@@ -58,9 +58,10 @@
     if (isValidCategories === true);
     if (!validate) return false;
     const userId = await findUserId(req);
+    
     const { id } = await BlogPost.create({ title, content, userId });
-     await createPostCategory(id, categoryIds);
- 
+    await createPostCategory(id, categoryIds);
+    
     return res.status(201).json({ id, title, content, userId });
   };
 
@@ -85,8 +86,6 @@
       ],
     });
 
-    console.log(blogPost.dataValues);
-
   if (blogPost === null) {
     return res.status(404).json({ message: 'Post does not exist' });
   }
@@ -98,13 +97,17 @@ const editBlogPost = async (req, res) => {
   try {
     const { id } = req.params;
     const { title, content, categoryIds } = req.body;
-    // const { id: userId } = req.user;
+    const userId = await findUserId(req);
+    const postId = await BlogPost.findByPk(id);
 
-    console.log(id);
-    const blogPost = await 
+   const blogPost = await 
     editBlogPostService
-    .editBlogPost({ id }, { title, content, categoryIds } /* { id: userId } */, res);
-    
+    .editBlogPost({ id }, { title, content, categoryIds }, res);
+
+  if (postId.id !== userId) {
+        return res.status(401).json({ message: 'Unauthorized user' });
+     }
+
     return res.status(200).json(blogPost);
   } catch (err) {
     res.status(500).json({
@@ -115,16 +118,22 @@ const editBlogPost = async (req, res) => {
 
 const deleteBlogPost = async (req, res) => {
   const { id } = req.params; 
-  const { id: userId } = req.user;  
-
-  const blogPost = await getPostById({ id });
-  
-  if (blogPost.userId !== userId) { 
-    return res.status(401).json({ message: 'Unauthorized user' }); 
-}
+  const postId = await BlogPost.findByPk(id);
+  const userId = await findUserId(req);
+  console.log('id', id);
+  console.log('postId', postId.id);
+  if (postId.id === null || id === null) {
+    return res.status(404).json({
+      message: 'Post does not exist',
+    });
+  }
+  console.log('userId', userId);
+ if (postId.id !== userId) {
+     return res.status(401).json({ message: 'Unauthorized user' }); 
+ }
 
   const deletedPost = await BlogPost.destroy({ where: { id } });
-  return res.status(204).json(deletedPost);
+  return res.status(204).json({ deletedPost });
 };
 
 module.exports = {
